frameworks/cognee/requirements.txt
cognee[ollama]
python-dotenv
asyncio
httpx
